+++ source/renderer/app/containers/wallet/WalletSettingsPage.js
@@ -20,7 +20,13 @@ export default class WalletSettingsPage extends Component<Props> {
   static defaultProps = { actions: null, stores: null };
 
   render() {
-    const { uiDialogs, walletSettings, app, wallets } = this.props.stores;
+    const {
+      uiDialogs,
+      walletSettings,
+      app,
+      wallets,
+      networkStatus,
+    } = this.props.stores;
     const activeWallet = wallets.active;
 
     // Guard against potential null values
@@ -64,6 +70,7 @@ export default class WalletSettingsPage extends Component<Props> {
         walletId={activeWallet.id}
         walletName={activeWallet.name}
         creationDate={creationDate}
+        isIncentivizedTestnet={networkStatus.isIncentivizedTestnet}
         isSubmitting={updateWalletRequest.isExecuting}
         isInvalid={
           updateWalletRequest.wasExecuted &&