+++ storybook/stories/WalletScreens-Settings.stories.js
@@ -139,6 +139,7 @@ export default () => {
       activeField={null}
       isInvalid={false}
       isSubmitting={false}
+      isIncentivizedTestnet={false}
       lastUpdatedField={null}
       nameValidator={() => true}
       onCancelEditing={() => {}}